# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LogiConnect Enterprise REST API
  description: |
    Enterprise internal communication and collaboration platform for logistics operations.
    Supports organizational hierarchy, granular RBAC, secure messaging, company broadcasts,
    document management, meeting coordination, and compliance audit logs.
  version: 1.0.0
  contact:
    name: LogiConnect Architecture Team
    email: architecture@logiconnect.internal
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
  - url: https://api.logiconnect.internal/api/v1
    description: Corporate Production Gateway
...
paths:
...
  /announcements:
    get:
      tags: [Announcements]
      summary: List Announcements Feed
      description: Paginated announcement feed visible to current authenticated employee based on active targeting scope.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
        - name: type
          in: query
          schema:
            type: string
            enum: [GENERAL, HR, OPERATIONS, SAFETY, POLICY, EMERGENCY, MAINTENANCE]
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, SCHEDULED, PUBLISHED, CANCELLED, ARCHIVED]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Announcements list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementPageEnvelope'
    post:
      tags: [Announcements]
      summary: Create Official Announcement
      description: Draft or schedule a new company, department, or team announcement.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnnouncementRequest'
      responses:
        '201':
          description: Announcement created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementDetailEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /announcements/{id}:
    get:
      tags: [Announcements]
      summary: Get Announcement Details (Records Read)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Announcement details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementDetailEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Announcements]
      summary: Update Announcement Draft
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAnnouncementRequest'
      responses:
        '200':
          description: Announcement updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementDetailEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /announcements/{id}/publish:
    post:
      tags: [Announcements]
      summary: Publish Announcement
      description: Transitions draft or scheduled announcement to PUBLISHED state with audit logging.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Announcement published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementDetailEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /announcements/{id}/schedule:
    post:
      tags: [Announcements]
      summary: Schedule Announcement
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleAnnouncementRequest'
      responses:
        '200':
          description: Announcement scheduled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementDetailEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /announcements/{id}/cancel:
    post:
      tags: [Announcements]
      summary: Cancel Announcement
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Announcement cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementDetailEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /announcements/{id}/archive:
    post:
      tags: [Announcements]
      summary: Archive Announcement
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Announcement archived successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementDetailEnvelope'
  /announcements/{id}/read:
    post:
      tags: [Announcements]
      summary: Mark Announcement as Read
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Read timestamp recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementReadEnvelope'
  /announcements/{id}/acknowledge:
    post:
      tags: [Announcements]
      summary: Acknowledge Mandatory Announcement
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Acknowledgement registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementReadEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /announcements/{id}/acknowledgements:
    get:
      tags: [Announcements]
      summary: Get Announcement Acknowledgement Report
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Acknowledgement compliance report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcknowledgementReportEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
  /announcements/{id}/statistics:
    get:
      tags: [Announcements]
      summary: Get Announcement Compliance Statistics (Alias)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Statistics report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcknowledgementReportEnvelope'

  # ==========================================================================
  # 10. DOCUMENTS API
  # ==========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageQuery:
      name: page
      in: query
      description: 0-indexed page number
      schema:
        type: integer
        default: 0
    SizeQuery:
      name: size
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        default: 20
        maximum: 100
...
  responses:
    BadRequest:
      description: Bad Request / Malformed Syntax
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Insufficient Role / Unauthorized Resource Access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Target Resource Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_FAILED"
            message:
              type: string
              example: "Provided payload violates validation constraints"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  rejectedValue:
                    type: string
                  issue:
                    type: string
        timestamp:
          type: string
          format: date-time
          example: "2026-08-31T12:00:00Z"
        path:
          type: string
          example: "/api/v1/employees"
    PaginationMetadata:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    CreateAnnouncementRequest:
      type: object
      required: [title, content, type, priority, targetType]
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [GENERAL, HR, OPERATIONS, SAFETY, POLICY, EMERGENCY, MAINTENANCE]
        priority:
          type: string
          enum: [NORMAL, HIGH, URGENT, EMERGENCY]
        targetType:
          type: string
          enum: [COMPANY, DEPARTMENT, TEAM]
        departmentId:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        requiresAcknowledgement:
          type: boolean
        scheduledAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    UpdateAnnouncementRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [GENERAL, HR, OPERATIONS, SAFETY, POLICY, EMERGENCY, MAINTENANCE]
        priority:
          type: string
          enum: [NORMAL, HIGH, URGENT, EMERGENCY]
        targetType:
          type: string
          enum: [COMPANY, DEPARTMENT, TEAM]
        departmentId:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        requiresAcknowledgement:
          type: boolean
        scheduledAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    ScheduleAnnouncementRequest:
      type: object
      required: [scheduledAt]
      properties:
        scheduledAt:
          type: string
          format: date-time
    AnnouncementDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        type:
          type: string
        status:
          type: string
        priority:
          type: string
        targetType:
          type: string
        departmentId:
          type: string
          format: uuid
        departmentName:
          type: string
        teamId:
          type: string
          format: uuid
        teamName:
          type: string
        requiresAcknowledgement:
          type: boolean
        scheduledAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        createdById:
          type: string
          format: uuid
        createdByName:
          type: string
        publishedById:
          type: string
          format: uuid
        publishedByName:
          type: string
        read:
          type: boolean
        readAt:
          type: string
          format: date-time
        acknowledged:
          type: boolean
        acknowledgedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AnnouncementPageEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AnnouncementDto'
            pagination:
              $ref: '#/components/schemas/PaginationMetadata'
        message:
          type: string
    AnnouncementDetailEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/AnnouncementDto'
        message:
          type: string
    AnnouncementReadEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            announcementId:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            read:
              type: boolean
            readAt:
              type: string
              format: date-time
            acknowledged:
              type: boolean
            acknowledgedAt:
              type: string
              format: date-time
        message:
          type: string
    EmployeeAcknowledgementStatusDto:
      type: object
      properties:
        employeeId:
          type: string
          format: uuid
        employeeCode:
          type: string
        fullName:
          type: string
        departmentName:
          type: string
        teamName:
          type: string
        read:
          type: boolean
        readAt:
          type: string
          format: date-time
        acknowledged:
          type: boolean
        acknowledgedAt:
          type: string
          format: date-time
    AcknowledgementReportEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            announcementId:
              type: string
              format: uuid
            title:
              type: string
            targetType:
              type: string
            departmentName:
              type: string
            teamName:
              type: string
            requiresAcknowledgement:
              type: boolean
            totalEligible:
              type: integer
            readCount:
              type: integer
            readPercentage:
              type: number
              format: double
            acknowledgedCount:
              type: integer
            acknowledgedPercentage:
              type: number
              format: double
            employeeStatuses:
              type: array
              items:
                $ref: '#/components/schemas/EmployeeAcknowledgementStatusDto'
        message:
          type: string

    # Document Schemas
...